fix(via): ws control flow
