--- src/builtin/ws/mod.rs
@@ -1,5 +1,0 @@
-mod handshake;
-mod message;
-
-pub use handshake::{Handshake, RequestContext, WebSocket, ws};
-pub use message::{CloseCode, Message};
